+++ specs/it.cpp
@@ -9,21 +9,21 @@ go_bandit([]() {
   describe("it", [&]() {
     std::function<void()> it_func;
     fake_reporter* reporter;
-    std::unique_ptr<bandit::detail::settings_t> settings;
+    std::unique_ptr<bandit::detail::controller_t> controller;
     std::unique_ptr<fake_context> context;
 
     before_each([&]() {
       reporter = new fake_reporter();
-      settings.reset(new bandit::detail::settings_t());
-      settings->set_reporter(reporter);
+      controller.reset(new bandit::detail::controller_t());
+      controller->set_reporter(reporter);
       context = std::unique_ptr<fake_context>(new fake_context());
-      settings->get_contexts().push_back(context.get());
+      controller->get_contexts().push_back(context.get());
 
-      settings->set_policy(new run_policy::always());
+      controller->set_policy(new run_policy::always());
     });
 
     auto call_it = [&]() {
-      it("my it", it_func, false, *settings);
+      it("my it", it_func, false, *controller);
     };
 
     it("tells the current context that execution has started", [&]() {
@@ -36,7 +36,7 @@ go_bandit([]() {
     });
 
     it("does not work without context", [&] {
-      settings->get_contexts().pop_back();
+      controller->get_contexts().pop_back();
       AssertThrows(bandit::detail::test_run_error, call_it());
       AssertThat(LastException<bandit::detail::test_run_error>().what(),
           Equals("'it' was called without surrounding 'describe'"));
@@ -85,7 +85,7 @@ go_bandit([]() {
 
         it("tells run_policy that we have a failing test", [&]() {
           call_it();
-          AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+          AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
         });
       });
 
@@ -106,7 +106,7 @@ go_bandit([]() {
 
         it("tells run_policy that we have a failing test", [&]() {
           call_it();
-          AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+          AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
         });
       });
 
@@ -127,7 +127,7 @@ go_bandit([]() {
 
         it("tells run_policy that we have a failing test", [&]() {
           call_it();
-          AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+          AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
         });
       });
 
@@ -149,7 +149,7 @@ go_bandit([]() {
 
         it("tells run_policy that we have a failing test", [&]() {
           call_it();
-          AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+          AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
         });
       });
 
@@ -170,7 +170,7 @@ go_bandit([]() {
 
         it("tells run_policy that we have a failing test", [&]() {
           call_it();
-          AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+          AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
         });
       });
 
@@ -191,7 +191,7 @@ go_bandit([]() {
 
         it("tells run_policy that we have a failing test", [&]() {
           call_it();
-          AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+          AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
         });
       });
     });
@@ -219,7 +219,7 @@ go_bandit([]() {
 
       it("tells run_policy that we have a failing test", [&]() {
         call_it();
-        AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+        AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
       });
     });
 
@@ -245,7 +245,7 @@ go_bandit([]() {
 
       it("tells run_policy that we have a failing test", [&]() {
         call_it();
-        AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+        AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
       });
     });
 
@@ -271,32 +271,32 @@ go_bandit([]() {
 
       it("tells run_policy that we have a failing test", [&]() {
         call_it();
-        AssertThat(settings->get_policy().has_encountered_failure(), IsTrue());
+        AssertThat(controller->get_policy().has_encountered_failure(), IsTrue());
       });
     });
 
     describe("it_skip", [&]() {
       it("tells reporter it's skipped", [&]() {
-        it_skip("my it", []() {}, *settings);
+        it_skip("my it", []() {}, *controller);
         AssertThat(reporter->call_log(), Has().Exactly(1).EqualTo("it_skip: my it"));
       });
 
       it("doesn't call function", [&]() {
         bool called = false;
-        it_skip("my it", [&]() { called = true; }, *settings);
+        it_skip("my it", [&]() { called = true; }, *controller);
         AssertThat(called, IsFalse());
       });
     });
 
     describe("xit", [&]() {
       it("tells reporter it's skipped", [&]() {
-        xit("my it", []() {}, *settings);
+        xit("my it", []() {}, *controller);
         AssertThat(reporter->call_log(), Has().Exactly(1).EqualTo("it_skip: my it"));
       });
 
       it("doesn't call function", [&]() {
         bool called = false;
-        xit("my it", [&]() { called = true; }, *settings);
+        xit("my it", [&]() { called = true; }, *controller);
         AssertThat(called, IsFalse());
       });
     });
@@ -305,7 +305,7 @@ go_bandit([]() {
       bool it_was_called;
 
       before_each([&]() {
-        settings->set_policy(new run_policy::never());
+        controller->set_policy(new run_policy::never());
         it_func = [&]() { it_was_called = true; };
         it_was_called = false;
       });
@@ -331,7 +331,7 @@ go_bandit([]() {
 
       describe("with a policy that says to skip this it", [&]() {
         before_each([&]() {
-          settings->set_policy(new run_policy::never());
+          controller->set_policy(new run_policy::never());
         });
 
         it("tells reporter it's skipped", [&]() {