+++ specs/run.cpp
@@ -12,24 +12,24 @@ go_bandit([]() {
     std::unique_ptr<bd::spec_registry> specs;
     std::unique_ptr<argv_helper> argv;
     fake_reporter* reporter;
-    std::unique_ptr<bd::settings_t> settings;
+    std::unique_ptr<bd::controller_t> controller;
 
     auto call_run = [&]() -> int {
       bd::options opt(argv->argc(), argv->argv());
-      return run(opt, *specs, *settings);
+      return run(opt, *specs, *controller);
     };
 
     before_each([&]() {
       specs.reset(new bd::spec_registry());
       reporter = new fake_reporter();
-      settings.reset(new bd::settings_t());
-      settings->set_reporter(reporter);
+      controller.reset(new bd::controller_t());
+      controller->set_reporter(reporter);
       argv.reset(new argv_helper({}));
     });
 
     it("pushes the global context on the context stack", [&]() {
       call_run();
-      AssertThat(settings->get_contexts(), Is().OfLength(1));
+      AssertThat(controller->get_contexts(), Is().OfLength(1));
     });
 
     describe("a successful test run", [&]() {